@@ -1,11 +1,10 @@
 name: accuracy_setting
 description: Get Location Accuracy Setting
-version: 0.0.1
-author:
-homepage:
+version: 2.0.0
+homepage: https://github.com/fuadarradhi/accuracy_setting
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
